requirements.txt
numpy #Numerical computations
opencv-python # Image processing
pillow  # Image handling
pytest